25: use int8 in negamax

+++ include/solver.h
@@ -2,6 +2,7 @@
 #define INCLUDE_SOLVER_H_
 
 #include <list>
+#include <cstdint>
 #include "../include/state.h"
 #include "../include/table.h"
 
@@ -9,7 +10,7 @@ class Solver {
  public:
   Solver();
   int negamax(State S) const;
-  int negamax(State S, int lowerBound, int upperBound) const;
+  int8_t negamax(State S, int8_t lowerBound, int8_t upperBound) const;
  private:
   Table table;
   std::list<int> moveOrder;
